restructure and remove caching tests
